Fix oUF_NotAuraTrack, allow multiple buffs to use the same icon
Need to track showing locally to self. Since we're always resetting it,
using an empty one local to the function will suffice.

Index by the icon table rather than the spellID for icons, and allow
multiple buffs to share an icon via icon.setTex.

Use that in the raidframes, because the number of turtle buffs is high.